Flip the catalog installation to the openwhisk-catalog repo
Currently, all the actions in openwhisk-catalog have been synchronized
with the ones in openwhisk. The test coverage of actions is equal to
openwhisk. We are technically able to switch from openwhisk/catalog to the
openwhisk-catalog.

WskBasicTests has been moved under whisk.core.cli.test, so that
whisk.core.entity.Subject can be accessed to create the fake user.