Flip the catalog installation to the openwhisk-catalog reop
Currently, all the actions in openwhisk-catalog have been synchronized
with the ones in openwhisk. The test coverage of actions is equal to
openwhisk. We are technically able to switch from openwhisk/catalog to the
openwhisk-catalog.

Since the package name 'util' is changed to 'utils', the test cases are
also change to use 'utils' instead of 'util'.