use ZooKeeper as persistent store for invokerId assignment (#2916)
Use ZooKeeper instead of Redis to persist mapping from invoker names to invoker ids and the counter for the next available id.

The basic algorithm for dynamic invoker id assignment is unchanged, just using a different persistent store whose durability is a better match for this purpose (we want the invoker id assignment information to be as durable as the kafka topics it is being used to assign).